Added CI .png output, more upscale comparisons and testing (yet to be added to main repo), copied all original data to new folders so each rip starts fresh, fixed various bugs with multiple areas not exporting correctly

+++ F3D.py
@@ -4,6 +4,7 @@ import math
 import time
 import struct
 import BinPNG
+import os
 
 #typedef struct {
 #  unsigned char	col[3];		/* diffuse light value (rgba) */
@@ -45,32 +46,49 @@ def ModelWrite(rom,ModelData,nameG,id,tdir):
 	'I':BinPNG.I
 	}
 	name = nameG/'model.inc.c'
-	textures = open(tdir/'textureNew.inc.c','w')
+	if os.path.isfile(tdir/'textureNew.inc.c'):
+		textures = open(tdir/'textureNew.inc.c','a')
+	else:
+		textures = open(tdir/'textureNew.inc.c','w')
 	f = open(name,'w')
 	f.write('#include "%s"\n'%('model.inc.h'))
 	for md in ModelData:
 		#textures
 		for t in md[3]:
 			if t[0]:
-				#textureptrs = raw ptr, bank ptr, length, width, height, imgtype, bitdepth, palette
+				#textureptrs = raw ptr, bank ptr, length, width, height, imgtype, bitdepth, palette, tile
 				if t in txt:
 					continue
 				texn = 'const u8 texture_%s_custom[]'%(id+hex(t[1]))
 				txt.append(t)
 				refs.append(texn)
-				if t[5]=='CI' or t[7]:
-					f.write('ALIGNED8 '+texn+' = {\n')
-					for i in range(t[2]):
-						h=rom[t[0]+i*2:t[0]+i*2+2]
-						f.write("0x{:02X},".format(int(h.hex(),16)))
-						if i%16==15:
-							f.write('\n')
-					f.write('};\n\n')
+				if t[5]=='CI':
+					texnp = 'const u8 texture_%s_custom_pal[]'%(id+hex(t[1]))
+					#export a include of a png file
+					textures.write('ALIGNED8 '+texn+' = {\n')
+					inc = "levels/"+tdir.parts[-1]+"/"
+					textures.write('#include "%s.inc.c"\n};\n'%(str(inc+(id+hex(t[1])+"_custom.%s%d"%(t[5].lower(),t[6])))))
+					#The palette
+					textures.write('ALIGNED8 '+texnp+' = {\n')
+					inc = "levels/"+tdir.parts[-1]+"/"
+					textures.write('#include "%s_pal.inc.c"\n};\n'%(str(inc+(id+hex(t[1])+"_custom.%s%d"%(t[5].lower(),t[6])))))
+					#export a png
+					bin = rom[t[0]:t[0]+t[2]*2+2]
+					pal = [rom[t[7][0]:t[7][0]+(2**t[6])*2],'rgba16'] #A palette is 2^bitdepth of CI * two bytes per pixel
+					png = BinPNG.MakeImage(str(tdir/(id+hex(t[1])+"_custom.%s"%(t[5].lower()+str(t[6])))))
+					png = ImgTypes[t[5]](t[3],t[4],t[6],pal,bin,png)
+					# f.write('ALIGNED8 '+texn+' = {\n')
+					# for i in range(t[2]):
+						# h=rom[t[0]+i*2:t[0]+i*2+2]
+						# f.write("0x{:02X},".format(int(h.hex(),16)))
+						# if i%16==15:
+							# f.write('\n')
+					# f.write('};\n\n')
 				else:
 					#export a include of a png file
 					textures.write('ALIGNED8 '+texn+' = {\n')
 					inc = "levels/"+tdir.parts[-1]+"/"
-					textures.write('#include "%s.inc.c"\n};'%(str(inc+(id+hex(t[1])+"_custom.rgba16"))))
+					textures.write('#include "%s.inc.c"\n};\n'%(str(inc+(id+hex(t[1])+"_custom.%s%d"%(t[5].lower(),t[6])))))
 					#export a png
 					bin = rom[t[0]:t[0]+t[2]*2+2]
 					png = BinPNG.MakeImage(str(tdir/(id+hex(t[1])+"_custom.%s"%(t[5].lower()+str(t[6])))))
@@ -178,7 +196,7 @@ def DecodeDL(rom,start,s,id):
 	#needs (ptr,length)
 	verts=[]
 	#needs (ptr,length)
-	textureptrs=[[0,0,0,0,0,0,0,0]]
+	textureptrs=[[0,0,0,0,0,0,0,0,0]]
 	#needs ptr
 	amb=[]
 	#neess ptr
@@ -231,25 +249,26 @@ def DecodeDL(rom,start,s,id):
 		#if a triangle is drawn and there is a texture, assume a new one is loaded next
 		elif(cmd[1][:8].uint==0xBF):
 			if textureptrs[-1][0]:
-				textureptrs.append([0,0,0,0,0,0,0,0])
-		#textureptrs = raw ptr, bank ptr, length, width, height, imgtype, bitdepth, palette
+				textureptrs.append([0,0,0,0,0,0,0,0,0])
+		#textureptrs = raw ptr, bank ptr, length, width, height, imgtype, bitdepth, palette, tile
 		#implementing a very naive alg because I'm lazy and no one hand writes stuff
 		#so I will just assume it follows nice structure, if you want to make it better then PR
 		#set tile
 		elif(cmd[1][:8].uint==0xf5):
 			tile = cmd[1][32:40].uint
-			if tile==0:
+			if tile!=7:
 				type=cmd[1][8:11].uint
 				textureptrs[-1][5]=types[type]
 				bpp=4*2**(cmd[1][11:13].uint)
 				textureptrs[-1][6]=bpp
+				textureptrs[-1][8]=tile
 		#tlut
 		elif(cmd[1][:8].uint==0xf0):
-			textureptrs[-1][7]=1
-			textureptrs.append([0,0,0,0,0,0,0])
+			tile = cmd[1][32:40].uint
+			if textureptrs[-1][8]==tile:
+				textureptrs[-1][7]=textureptrs[-1][:2]
 		#set tile size
 		elif(cmd[1][:8].uint==0xf2):
-			tile = cmd[1][32:40].uint
 			f2 = (lambda x: (x>>2)+1)
 			textureptrs[-1][3] = f2(cmd[1][40:52].uint)
 			textureptrs[-1][4] = f2(cmd[1][52:64].uint)
